Add Sdl class for OS-specific DLL import
A new Sdl class was added in the Imports namespace of our SDL2 .NET
library to handle the OS-specific DLL imports.

It uses a KeyValuePair list to map different OS to their corresponding
DLL file names.

The class also provides a DllImportResolver to load the right DLL based
on the running OS.

Signed-off-by: Victor Matia <vmatir@gmail.com>