Add CPU and SIMD utilities and related imports
This commit introduces utilities for gathering CPU information and SIMD
functionalities.

New `CpuInfo` and `Simd` classes have been added to
`Vmr.Sdl2.Net.Utilities` and their related imports are included in
`Imports`.

This addition will enhance the ability of the software to interact and
gain information from the CPU and perform SIMD operations.

Signed-off-by: Victor Matia <vmatir@gmail.com>