Add version mismatch handling mechanism
The commit introduces a delegate for handling version mismatches in the
NativeLibraryInformation class and adds a new optional parameter to the
Application constructor to hold the version mismatch handler.

If provided, it is invoked with the expected and actual versions.

In addition, a new ExpectedVersion property has been defined for
internal use.

Reformatted the `Sdl.cs` file to add a new line for consistency with the
rest of the project.

Signed-off-by: Victor Matia <vmatir@gmail.com>