Add new graphics and colors related classes and enums
Several new files are created to provide constructs related to graphics
and colors in the Vmr.Sdl2.Net library.

These include enums for ArrayOrder, BitmapOrder, PackedLayout,
PackedOrder, and classes for ColorLoss, ColorMasks, ColorShift,
Palettes and more.

The addition of these classes and enums will improve the handling and
manipulation of graphics and colors in the library.

Signed-off-by: Victor Matia <vmatir@gmail.com>