Modified classes and structs to allow data comparison and serialization
Several classes and structures in different namespaces have been
enhanced to implement the IEquatable interface, which provides a
stronger model for comparing objects of their types.

This also includes attaching the Serializable attribute to these classes
and structures to support serialization.

Additionally, the ToString method is implemented to enhance value
readability.

Signed-off-by: Victor Matia <vmatir@gmail.com>