Add Code of Conduct and update documentation links
Introduced a new comprehensive Code of Conduct to ensure a respectful,
inclusive and harassment-free environment for all project participants.

This Code of Conduct was adapted from the Contributor Covenant and has
been added to the project's README and CONTRIBUTING files.

Updated the .csproj file to reflect the addition and changes to the
documentation files and packaging fixes.

Signed-off-by: Victor Matia <vmatir@gmail.com>