Update battery remaining time representation
Changed the battery remaining time attribute type from int to TimeSpan
in the BatteryLife class to better represent the time duration.

This update affects the BatteryLife properties and associated methods,
providing a more accurate duration representation.

The change also includes an adapter in the PowerInformation class to
convert the seconds into TimeSpan.

Signed-off-by: Victor Matia <vmatir@gmail.com>