Change DebugType in Release configuration
The DebugType in the Release configuration of the Vmr.Sdl2.Net project
has been modified.

Previously set to "none", it has been changed to "pdbonly" to allow
generation of .pdb symbol files even in release mode.

This facilitates debugging without affecting the code's performance, and
allows creating a release snupkg.

Signed-off-by: Victor Matia <vmatir@gmail.com>