Massive rewrite to change the error system.
Now errors use exceptions, as expected for C#.

Created multiple exceptions to manage the library errors.

Modified the `Application` class to manage the initialization, loading,
updating and main loop logic, making it possible to use as a base for
programs.

Changed the `MainLoop` class to `Events` to better represent its new
functionality.

Signed-off-by: Victor Matia <vmatir@gmail.com>